chore(hlapi): add some GPU test for FheUint
Tests are not complete yet, but its the first step to get there
